Added prod docker setup. Polished some configs. Minor tweaks and improvements.

+++ scripts/reset_local_data.sh
@@ -4,7 +4,14 @@
 
 _env="$1"
 file_path="$2"
-_dkc() { docker-compose --log-level ERROR "$@"; }
+_dkc() {
+
+  if [[ $_env == "prod" ]]; then
+    docker-compose --log-level ERROR -f docker-compose.prod.yml "$@";
+  else
+    docker-compose --log-level ERROR "$@";
+  fi
+}
 _dkc-run() { _dkc run -e LOGGING_LEVEL_CONSOLE=WARNING --rm "$@"; }
 _dkc-run-django() {
   if [[ $_env == "local" ]]; then
@@ -21,18 +28,21 @@ else
   echo "Using 'django' docker container"
 fi
 
-echo "Recreating blank database 'atss'"
-_dkc restart postgres
-while ! docker-compose --log-level ERROR run --no-deps postgres psql -h postgres -U postgres -c 'select 1;' > /dev/null 2>&1; do
-    echo "Waiting for postgres initialization..."
-    sleep 1
-done
 
-_dkc exec postgres dropdb --if-exists -h postgres -U postgres atss
-echo "Dropped 'atss' database."
+if [[ $_env != "prod" ]]; then
+  echo "Recreating blank database 'atss'"
+  _dkc restart postgres
+  while ! docker-compose --log-level ERROR run --no-deps postgres psql -h postgres -U postgres -c 'select 1;' > /dev/null 2>&1; do
+      echo "Waiting for postgres initialization..."
+      sleep 1
+  done
 
-_dkc exec postgres createdb -h postgres -U postgres atss
-echo "Re-created 'atss' database."
+  _dkc exec postgres dropdb --if-exists -h postgres -U postgres atss
+  echo "Dropped 'atss' database."
+
+  _dkc exec postgres createdb -h postgres -U postgres atss
+  echo "Re-created 'atss' database."
+fi
 
 echo "Rebuilding database schema..."
 _dkc-run-django python3 manage.py migrate
@@ -41,7 +51,7 @@ echo "Loading fixtures..."
 _dkc-run-django python3 manage.py loaddata init
 
 echo "Loading initial data from $file_path..."
-_dkc-run-django python3 manage.py init_data --file_path=$file_path -v 3
+_dkc-run-django python3 manage.py init_data --file_path=$file_path
 
 _end=$(date +%s)
 echo "Database reset complete! (took $((_end-_start))s)"
